backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
xarray>=2024.1.0
cfgrib>=0.9.10
eccodes>=1.6.0
rasterio>=1.3.0
rio-tiler>=6.2.0
numpy>=1.26.0
Pillow>=10.1.0
python-dotenv>=1.0.0
